moves coroutines to specific class
